Move top nav logic for different sections of the app to their own reducer
